+++ packages/bpk-docs/src/pages/DatepickerPage/DatepickerPage.js
@@ -1,8 +1,10 @@
 import BpkParagraph from 'bpk-component-paragraph';
 import React from 'react';
 import BpkDatepicker from 'bpk-component-datepicker';
+import BpkRouterLink from 'bpk-component-router-link';
 
 import datepickerReadme from 'bpk-component-datepicker/readme.md';
+import * as ROUTES from './../../constants/routes';
 import DocsPageBuilder from './../../components/DocsPageBuilder';
 import format from '../../../../bpk-component-calendar/node_modules/date-fns/format';
 import { weekDays } from '../../../../bpk-component-calendar/test-utils';
@@ -20,7 +22,7 @@ const components = [
         Default calendar, input and popover configuration.
       </BpkParagraph>,
       <BpkParagraph>
-        See the BpkCalendar documentation for more examples.
+        See the <BpkRouterLink to={ROUTES.CALENDAR}>BpkCalendar documentation</BpkRouterLink> for more examples.
       </BpkParagraph>,
     ],
     examples: [
@@ -42,8 +44,9 @@ const DatepickerPage = () => <DocsPageBuilder
   title="Datepicker"
   blurb={[
     <BpkParagraph>
-      The datepicker is an input control that allows the user to pick a single date. It embeds a calendar
-      component in a popover.
+      The datepicker is an input control that allows the user to pick a single date. It embeds
+      a <BpkRouterLink to={ROUTES.CALENDAR}>calendar component</BpkRouterLink> in
+      a <BpkRouterLink to={ROUTES.POPOVERS}>popover</BpkRouterLink>.
     </BpkParagraph>,
   ]}
   components={components}